+++ src/images/repositories/images.respository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateImageDto } from '../dto/create-image.dto';
 import { ImageEntity } from '../entities/image.entity';
@@ -19,17 +19,29 @@ export class ImageRepository {
   }
 
   async findOne(id: string): Promise<ImageEntity> {
-    return this.prisma.image.findUnique({
+    const findImage = await this.prisma.image.findUnique({
       where: {
         id,
       },
     });
+    if (!findImage) {
+      throw new NotFoundException('Image not found');
+    }
+    return findImage;
   }
 
   async update(
     id: string,
     updateImageDto: UpdateImageDto,
   ): Promise<ImageEntity> {
+    const findImage = await this.prisma.image.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!findImage) {
+      throw new NotFoundException('Image not found');
+    }
     return this.prisma.image.update({
       where: {
         id,
@@ -39,6 +51,14 @@ export class ImageRepository {
   }
 
   async remove(id: string): Promise<ImageEntity> {
+    const findImage = await this.prisma.image.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!findImage) {
+      throw new NotFoundException('Image not found');
+    }
     return this.prisma.image.delete({
       where: {
         id,
